jwt authentication and other fixes
