+++ package.json
@@ -23,7 +23,6 @@
     "body-parser": "^1.17.1",
     "core-js": "^2.4.1",
     "express": "^4.15.2",
-    "jsonwebtoken": "^8.1.0",
     "lodash": "^4.17.4",
     "mongoose": "^4.10.0",
     "rxjs": "^5.1.0",
