+++ src/app/app.module.ts
@@ -32,14 +32,10 @@ import { JoinComponent } from './join';
 
 import { ContactusService } from './services/contactus.service';
 import { UserService } from './services/user.service';
-import { ErrorService } from './services/error.service';
-
-import { AuthGuard } from './auth.guard';
 
 import { ROUTES } from './app.route';
 import { LoginComponent } from './login/login.component';
 import { SignupComponent } from './signup/signup.component';
-import { ProfileComponent } from './profile/profile.component';
 
 @NgModule({
   declarations: [
@@ -62,8 +58,7 @@ import { ProfileComponent } from './profile/profile.component';
     ContactusComponent,
     JoinComponent,
     LoginComponent,
-    SignupComponent,
-    ProfileComponent
+    SignupComponent
   ],
   imports: [
     BrowserModule,
@@ -72,7 +67,7 @@ import { ProfileComponent } from './profile/profile.component';
     HttpModule,
     RouterModule.forRoot( ROUTES, { useHash: true, preloadingStrategy: PreloadAllModules })
   ],
-  providers: [ ErrorService, AuthGuard, DefaultRequestOptions, ContactusService, UserService],
+  providers: [ DefaultRequestOptions, ContactusService, UserService],
   bootstrap: [AppComponent]
 })
 export class AppModule { }
