--- src/app/auth.guard.ts
@@ -1,21 +1,0 @@
-import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot } from '@angular/router';
-import { Observable } from 'rxjs/Observable';
-import { UserService } from './services/user.service';
-
-@Injectable()
-export class AuthGuard implements CanActivate {
-
-  constructor(private UserService: UserService) {
-  }
-
-  canActivate(
-    next: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean> | Promise<boolean> | boolean {
-      if ( this.UserService.getUser()) {
-        return true;
-      }else {
-        return false;
-      }
-  }
-}
