+++ src/app/login/login.component.html
@@ -25,10 +25,6 @@
             </div>
             <div class="col-md-4">
                 <input type="submit" id="btn-login" class="btn btn-custom btn-lg btn-block" value="Log in" [disabled]="!loginForm.form.valid">
-                <br>
-                <a href="#signup" id="btn-login" class="btn btn-custom btn-lg btn-block btn-success">sign up</a>
-                <br>
-                <p>{{errorMessage}}</p>
             </div>
             <div class="col-md-4">
             </div>
