--- src/app/profile/profile.component.css
@@ -1,3 +1,0 @@
-.buffer{
-    margin-top:60px;
-}
