+++ src/app/shared/footer/footer.component.html
@@ -4,17 +4,17 @@
 
     <h3>Join Us</h3>
 
-    <!-- <form class="mc-trial row">
+    <form class="mc-trial row">
       <div class="form-group col-md-3 col-md-offset-2 col-sm-4">
         <div class=" controls">
           <input name="name" placeholder="Enter Your Name" class="form-control" type="text">
         </div>
-      </div>
+      </div><!-- End email input -->
       <div class="form-group col-md-3 col-sm-4">
         <div class=" controls">
           <input name="EMAIL" placeholder="Enter Your email" class="form-control" type="email">
         </div>
-      </div>
+      </div><!-- End email input -->
 
       <div class="col-md-2 col-sm-4">
         <p>
@@ -22,7 +22,7 @@
             Join now <i class="fa fa-arrow-right"></i></button>
         </p>
       </div>
-    </form>-->
+    </form><!-- End newsletter-form -->
     <ul class="social-links">
       <li><a href="https://twitter.com/ng_pakistan" target="_blank"><i class="fa fa-twitter fa-fw"></i></a></li>
       <li><a href="https://www.facebook.com/ngPakistan/" target="_blank"><i class="fa fa-facebook fa-fw"></i></a></li>
