+++ src/app/shared/navigation/navigation.component.ts
@@ -1,6 +1,4 @@
 import { Component, OnInit } from '@angular/core';
-import { Router } from '@angular/router';
-import { UserService } from '../../services/user.service';
 
 @Component({
   selector: 'app-navigation',
@@ -9,21 +7,9 @@ import { UserService } from '../../services/user.service';
 })
 export class NavigationComponent implements OnInit {
 
-  nameHover= 'btn-trial';
+  constructor() { }
 
-  constructor(private router: Router, private UserService: UserService) { }
-  showOptions() {
-    this.nameHover = 'btn-trial';
-  }
-  hideOptions(){
-    this.nameHover = '';
-  }
   ngOnInit() {
-  }
-
-  logout() {
-    this.UserService.clearUser();
-
   }
 
 }
