+++ src/app/signup/signup.component.ts
@@ -1,6 +1,5 @@
 import { Component, OnInit } from '@angular/core';
 import { user } from '../model/user.interface';
-import { Router } from '@angular/router';
 
 import { UserService } from '../services/user.service';
 
@@ -28,7 +27,8 @@ export class SignupComponent implements OnInit {
     password: ''
   };
 
-  constructor(private UserService: UserService, private Router: Router) { }
+  
+  constructor(private UserService: UserService) { }
 
   ngOnInit() {
     console.log(this.User);
@@ -36,9 +36,7 @@ export class SignupComponent implements OnInit {
 
   signup(){
     this.UserService.save(this.User).subscribe( data => {
-      if (data.data) {
-        this.Router.navigate(['login']);
-      }
+      console.log(data);
     });
   }
 }
